requirements.txt
quickbe==0.0.6
GitPython
cryptography
schedule
